+++ C++/kd-tree/kdTreeNlogn.h
@@ -1408,7 +1408,7 @@ public:
    * numNeighbors - the number M of nearest neighbors to attempt to find
    */
 public:
-  void findNearestNeighbors(forward_list< pair<double, KdNode<K>*> >& neighbors,
+  void findNearestNeighbors(forward_list< pair<cpp_int, KdNode<K>*> >& neighbors,
                             vector<K> const& query,
                             signed_size_t const numNeighbors) {
     
@@ -1428,7 +1428,7 @@ public:
    * enable - a vector that specifies the dimensions for which to test distance
    */
 public:
-  void findNearestNeighbors(forward_list< pair<double, KdNode<K>*> >& neighbors,
+  void findNearestNeighbors(forward_list< pair<cpp_int, KdNode<K>*> >& neighbors,
                             vector<K> const& query,
                             signed_size_t const numNeighbors,
                             vector<bool> const& enable) {
@@ -1451,8 +1451,8 @@ public:
    * of the k-d tree for access to the KdNodes via the lists. Hence, this function is not static.
    */
 public:
-  void verifyNearestNeighbors(forward_list< pair<double, KdNode<K>*> >& neighborsFast,
-                              forward_list< pair<double, KdNode<K>*> >& neighborsSlow) const {
+  void verifyNearestNeighbors(forward_list< pair<cpp_int, KdNode<K>*> >& neighborsFast,
+                              forward_list< pair<cpp_int, KdNode<K>*> >& neighborsSlow) const {
     
     if (root != nullptr) {
       root->verifyNearestNeighbors(neighborsFast, neighborsSlow);
@@ -1468,21 +1468,21 @@ public:
    * query - a vector that contains the query point coordinates
    * maxNodes - the maximum number of nodes to maintain on the heap
    * 
-   * returns: a sorted forward list of (double, KdNode*) pairs
+   * returns: a sorted forward list of (cpp_int, KdNode*) pairs
    *
    * Because this function does not access the k-d tree, it could be static.
    * However, calling it as a static function requires speicification of a
    * type, so calling it as a non-static function is less cumbersome.
    */
 public:
-  forward_list<pair<double, KdNode<K>*>> sortByDistance(list<KdNode<K>*> const& kdList,
-                                                        vector<K> const& query,
-                                                        signed_size_t const& maxNodes) {
+  forward_list<pair<cpp_int, KdNode<K>*>> sortByDistance(list<KdNode<K>*> const& kdList,
+                                                         vector<K> const& query,
+                                                         signed_size_t const& maxNodes) {
 
     if (root != nullptr) {
       return root->sortByDistance(kdList, query, maxNodes);
     } else {
-      forward_list<pair<double, KdNode<K>*>> sortedList;
+      forward_list<pair<cpp_int, KdNode<K>*>> sortedList;
       return sortedList;
     }
   }
@@ -1511,8 +1511,8 @@ public:
    * maximumSubmitDepth - the maximum tree depth at which a child task may be launched
    */
 public:
-  void findReverseNearestNeighbors(vector< forward_list< pair<double, KdNode<K>*> > >& nn,
-                                   vector< forward_list< pair<double, KdNode<K>*> > >& rnn,
+  void findReverseNearestNeighbors(vector< forward_list< pair<cpp_int, KdNode<K>*> > >& nn,
+                                   vector< forward_list< pair<cpp_int, KdNode<K>*> > >& rnn,
                                    vector<mutex>& mutexes,
                                    signed_size_t const numDimensions,
                                    signed_size_t const numNeighbors,
@@ -1548,8 +1548,8 @@ public:
    * enable - a vector that specifies the dimensions for which to test distance
    */
 public:
-  void findReverseNearestNeighbors(vector< forward_list< pair<double, KdNode<K>*> >* >& nn,
-                                   vector< forward_list< pair<double, KdNode<K>*> >* >& rnn,
+  void findReverseNearestNeighbors(vector< forward_list< pair<cpp_int, KdNode<K>*> >* >& nn,
+                                   vector< forward_list< pair<cpp_int, KdNode<K>*> >* >& rnn,
                                    vector<mutex>& mutexes,
                                    signed_size_t const numDimensions,
                                    signed_size_t const numNeighbors,
@@ -1575,8 +1575,8 @@ public:
    * of the k-d tree for access to the KdNodes via the vectors. Hence, this function is not static.
    */
 public:
-  void verifyReverseNeighbors(vector< forward_list< pair<double, KdNode<K>*> > >& nn,
-                              vector< forward_list< pair<double, KdNode<K>*> > >& rnn,
+  void verifyReverseNeighbors(vector< forward_list< pair<cpp_int, KdNode<K>*> > >& nn,
+                              vector< forward_list< pair<cpp_int, KdNode<K>*> > >& rnn,
                               signed_size_t const numberOfNodes) {
 
     if (root != nullptr) {
@@ -1595,7 +1595,7 @@ public:
    * of the k-d tree for access to the KdNodes via the vector. Hence, this function is not static.
    */
 public:
-  void calculateMeanStd(vector< forward_list< pair<double, KdNode<K>*> > >& vec,
+  void calculateMeanStd(vector< forward_list< pair<cpp_int, KdNode<K>*> > >& vec,
                         double& meanSize,
                         double& stdSize,
                         double& meanDist,
@@ -1617,7 +1617,7 @@ public:
    * of the k-d tree for access to the KdNodes via the vector. Hence, this function is not static.
    */
 public:
-  size_t nonEmptyLists(vector< forward_list< pair<double, KdNode<K>*> > >& vec) const {
+  size_t nonEmptyLists(vector< forward_list< pair<cpp_int, KdNode<K>*> > >& vec) const {
 
     if (root != nullptr) {
       return root->nonEmptyLists(vec);
@@ -1638,7 +1638,7 @@ public:
    * numNeighbors - the number M of nearest neighbors to find
    */
 public:
-  void bruteNearestNeighbors(forward_list< pair<double, KdNode<K>*> >& neighbors,
+  void bruteNearestNeighbors(forward_list< pair<cpp_int, KdNode<K>*> >& neighbors,
                              vector<K> const& query,
                              signed_size_t const numNeighbors) {
     
@@ -1715,7 +1715,7 @@ public:
    *
    * Calling parameters:
    *
-   * regionList - a list of (double, KdNode*) pairs
+   * regionList - a list of (cpp_int, KdNode*) pairs
    * maximumNumberOfNodesToPrint - the maximum number of KdNodes to print
    * numDimensions - the number of dimensions
    *
@@ -1724,7 +1724,7 @@ public:
    * type, so calling it as a non-static function is less cumbersome.
    */
 public:
-  void printTuples(forward_list<pair<double, KdNode<K>*>> const& regionList,
+  void printTuples(forward_list<pair<cpp_int, KdNode<K>*>> const& regionList,
                    signed_size_t const maximumNumberOfNodesToPrint,
                    signed_size_t const numDimensions) const {
