+++ C++/kd-tree/test_kdtreedynamic.stats.cpp
@@ -531,7 +531,7 @@ int main(int argc, char** argv) {
 
       // Find numNeighbors nearest neighbors to each coordinate.
       if (neighbors) {
-        forward_list< pair<double, KdNode<kdKey_t>*> > neighborList;
+        forward_list< pair<cpp_int, KdNode<kdKey_t>*> > neighborList;
         vector<kdKey_t> query(coordinates[0].size());
         auto beginTime = steady_clock::now();
         for (size_t i = 0; i < coordinates.size(); ++i) {
@@ -606,7 +606,7 @@ int main(int argc, char** argv) {
 
     // Find numNeighbors nearest neighbors to each coordinate.
     if (neighbors) {
-      forward_list< pair<double, KdNode<kdKey_t>*> > neighborList;
+      forward_list< pair<cpp_int, KdNode<kdKey_t>*> > neighborList;
       vector<kdKey_t> query(coordinates[0].size());
       auto beginTime = steady_clock::now();
       for (size_t i = 0; i < coordinates.size(); ++i) {
